stress-ping-sock: cast return from getpid() to match type of echo.id
Explicitly cast (and show we are type truncating) return value from
getpid() to clear up implicity cast warning.

+++ stress-ping-sock.c
@@ -76,7 +76,7 @@ static int stress_ping_sock(const stress_args_t *args)
 	icmp_hdr = (struct icmphdr *)buf;
 	(void)memset(icmp_hdr, 0, sizeof(*icmp_hdr));
 	icmp_hdr->type = ICMP_ECHO;
-	icmp_hdr->un.echo.id = getpid();	/* some unique ID */
+	icmp_hdr->un.echo.id = (uint16_t)getpid();	/* some unique ID */
 	icmp_hdr->un.echo.sequence = 1;
 
 	rand_port = 1024 + (stress_mwc16() % (65535 - 1024));
